Eric Sunshine eliminated a batch of compilation warnings in OpenGL renderer.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@27298 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/opengl/gl_r2t_ext_fb_o.cpp
@@ -74,7 +74,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(gl3d)
   {
     GLint currentFB;
     glGetIntegerv (GL_FRAMEBUFFER_BINDING_EXT, &currentFB);
-    if (currentFB == framebuffer)
+    if (currentFB == (GLint)framebuffer)
       ext->glBindFramebufferEXT (GL_FRAMEBUFFER_EXT, 0);
 
     if (depthRB != 0)
